ref(main): Use elegant_departure to cleanly shutdown
The taskbroker has two shutdown cases currently:
1. A signal is received of some kind (SIGINT etc.)
2. One of the components panics/shuts down

In the first case, shut down the guard that all the components are awaiting,
which will cause them to execute.

In the second case, when the tokio task finishes, the guard being dropped
will trigger a shutdown on the other guards. Log the error/completion of
the original component and then wait for all the other components to shut down.