fix(errors): Handle panics/errs in components (#37)
The code wasn't correctly handling the different panic and error cases with the components.

Originally, if a component panicked, the panic would be handled by tokio and propagated to the
`log_task_completion` function correctly. However there is another case, which is that one of the
components finishes its `.await` with an error. This is particularly possible for the consumer,
which has multiple green threads, any of which could panic and propagate an error up the chain.

Change the consumer to correctly handle this case, and propagate an error if one of its green
threads panics or in returns an error.

Also change the main function to correctly log the error in this case. The shutdown logic doesn't
have to change since it already was triggered when one of the components finished executing.