fix(build): Changes required to compile binary
In order to compile and run the binary in separate steps, the VERSION global had to be changed to
not use a build ENV variable, which isn't available after the compilation is complete.

Update the code to write the VERSION to a file that can then be picked up by the executable when it
runs.