fix: Fix processing deadlines not being reset (#56)
When tasks had retry state and hadn't used all of their retries, the
upkeep component would not reset the processing deadlines because of
incorrect logic. When adding tests for these scenarios I realized that
we don't need to treat the 'has retries' and 'no retries' scenarios
differently and could simplify inflight activation logic.