ref(build): Add GHA to build image for sentry dev and prod (#54)
This adds 3 files for the use of generating docker images for the taskbroker.

It adds a cloudbuild.yaml. In GCB, this file is used to build an image that is stored in the 
Artifact Registry.

It adds an image.yaml file, which is responsible for building and pushing the `latest` container
for use in sentry devservices.

Finally, it adds a craft file, for legacy reasons.

It also makes a few small changes to how the VERSION file is generated, to work with how the
images are generated.