feat(k8s): Add basic k8s config files (#91)
This adds two k8s config files: one to create the storage class used for sqlite, and another to run
the taskbroker service itself.

The service currently uses a specific cloud image. This should be fixed in the future to be locked
to the latest build of the image, instead of a specific commit.

There are some rough instructions in the README on how to apply this k8s config to a sandbox for
testing.