feat(kafka): Create the topics if they don't exist (#102)
In dev situations, it is possible for the taskbroker to be started without the necessary topics existing locally. Add a function that will create the topics if they don't exist. This function can be skipped with a config setting.

This is necessary for a couple reasons: The taskbroker runs independently of sentry itself, in the same way snuba/relay run independently. That means it should be possible to start taskbroker without needing to start the sentry monolith. These types of applications are managed using the devservices commands (vs. devserver, which starts the monolith).

However, devservices has no mechanism to run commands on the dependent services, which is necessary in the taskbroker case to initialize the kafka topics.

In order to have taskbroker be independent of the sentry monolith, and to not have it require extra commands to bootstrap after running devservices up, it was necessary to create the topics manually when the service starts.

Note that this is also what both snuba and sentry do when they start up