feat: Add a Docker container for taskworkers
When testing in sandboxes, it's helpful to have a worker image that can be scaled up and down
easily, and that can be configured with different failure modes. Add a new docker image for a simple
worker that can be set to fail randomly a certain percentage of times.

The failure can be an actual task failure, or simply never returning a status for a task, causing it
to "timeout".