Replace DATETIME columns with INTEGER
SQlite doesn't have a specific storage type for datetime values and uses
TEXT instead. This requires careful use of `datetime()` functions on
value. By switching to INTEGER columns we won't get burned by future
changes to datetime logic if we forget a function. It should also result
in smaller database sizes.

I've also made `remove_at` required as we should not store an activation
that cannot be purged in the future.

Fixes #111