feat(expires): Track task expiry explicitly
Currently if a task has an expiry set, that expiry date is tracked through the remove_at column.
However that column is also used to track tasks that aren't being processed because of worker
failures.

The issue with that is that expired tasks should be handled differently from tasks that are not
being processed. Expired tasks can be removed immediately as soon as the expiry has passed,
regardless of any other tasks. Tasks that aren't being processed might be in that state because the
workers aren't running, so they can't be removed until other tasks start being completed.

Add an explicit column for expires_at, that is only populated if the task has an expiry. Also update
the upkeep thread to mark expired tasks for removal.

While in theory a check could be added to the consumer to avoid even ingesting tasks that are
expired, letting the upkeep thread remove them means the tasks get properly discarded/retried/dlq.