fix(metrics): Update metrics on processing and execution times (#204)
This is related to the bugs around task processing deadlines. The metric to track task execution
time was not accurate since it was ignoring the case when the processing deadline had passed.

This also tracks how often a task had its processing count incremented, not just the times the task
was discarded after exceeding the increment.