feat: Allow fetching multiple tasks at once
Change the store to allow fetching multiple pending tasks at once. This is the first step towards
being able to batch fetch tasks from the worker.

This adds a `get_pending_activations` function, and points the existing function to call that
function with a limit of 1. In the future new endpoints can leverage the pending activations
function directly. That function accepts the namespaces that are being requested, and the number of
pending tasks that can be returned.