fix(upkeep): Add a semaphore to prevent concurrent upkeep/vacuum
In a situtation where the broker is overloaded, and the upkeep/vacuum are taking longer than a
single interval to run, the old functionality would simply continue calling the function. If the
upkeep was taking 30 seconds to run, and there was a 1 second interval, that means there would be 30
concurrent upkeeps all running at the same time, and all effectively doing the same thing.

In the case of an incident like INC-1158, this would compound the issue the broker was facing: it
was already under high load, and this would simply keep piling on more and more load, as well as
potentially consuming a lot of memory keeping all the stack traces in memory.

Add a simple semaphore to limit the number of concurrent upkeeps/vacuums that can be running at once
to avoid this compounding issue. Hopefully this will help prevent high load causing incidents.