fix(gocd): Use exclude_regions (#366)
`include_regions` is a little non-intuitive, it's not including to an empty list but including to a default list (all `getsentry` regions).  `include_regions` is used for adding `control` to specific pipelines as we run a subset of our services for `control`.

To limit to regions we want to deploy to we need to use `exclude_regions` which removes the provided list of regions from the default list.