feat(sqlite): Split activation blobs to a separate DB
This change creates a new SQLite DB that stores the TaskActivation protobuf, keyed by ID. This table
is used to hydrate the protobuf when it is needed in the taskbroker. The original table remains,
however it won't contain the protobuf data. In order to ensure that processing could happen without
needing the protobuf itself, the on_attempts_exceeded value needed to be pulled out of the protobuf
and stored in the table directly.

There are now extra functions for fetching just the activation data, `get_activation_blob` and
`get_activation_raw_blob`. These fetch the decoded and encoded versions of the protobuf
respectively.

There is also a new upkeep function that deletes all the blobs that don't have a corresponding ID in
the meta table. Upkeep works essentially unchanged for the original table. Once a row is deleted
from the meta table, the corresponding blob is considered orphaned. The upkeep function deletes all
such blobs from the table that are older than the given delay_ms.

All of this is gated behind a config option that determines whether just the old table is used or
the new table is used. There are four modes for this migration:
1. Write only to the original table, and read only from the original table (same as before)
2. Write to both tables, but read from the original table.
3. Write to both tables, and read from the new table.
4. Write only to the new table and read from the new table.

There will be separate PRs moving the code through the different states, which allows for easy
rollbacks in case of problems. Since the new column added to the original table is nullable, this PR
can be rolled back safely as well.