feat: Add ability to perform a full vacuum on startup (#442)
* feat: Add ability to perform a full vacuum on startup

Should a broker run out of storage space, it is difficult for us to
force a manual vacuum as if the container is running the application
will have a write lock on the database. With these changes we can enable
a configuration via k8s that will force the broker to do a vacuum during
startup. This should help mitigate out of storage space issues in the
future.

Refs #440

* make format

* Use info! instead of println!