feat: Add latency metric from upkeep (#449)
* feat: Add latency metric during upkeep

The latency metric we generate from the GRPC endpoint is useful until
there are no workers available. When that happens we lose visibility
into how far behind a broker is. By generating a lag metric from upkeep
we can continue to emit import metrics when all workers are absent and
when ingestion is paused from backpressure.

Refs #448

* Move method

* Use milliseconds instead.

* Add another test.

* Remove print