feat(v2): Postgres storage adapter
This adds a postgres storage adapter for the taskbroker, as well as providing a way to choose
between the adapters in the configuration. This adapter will also work with AlloyDB.

In postgres, the keyword `offset` is reserved, so that column is called `kafka_offset` in the PG
tables and converted to `offset`.

The tests were updated to run with both the SQLite and Postgres adapter using the rstest crate. The
`create_test_store` function was updated to be the standard for all tests, and to allow choosing
between a SQLite and Postgres DB.

A `remove_db` function was added to the trait and the existing adapters, since the tests create a
unique PG database on every run that should be cleaned up.

The `create_test_store` function was updated to be the standard for all tests, and to allow choosing
between an SQLite and Postgres DB.